feat: move aria live-region to helpers and add to dom early
Add the ARIA live-region as a global helper, and append it to the body
earlier on during page load.
